Add sign in suggestion on auto-save failure
