+++ src/app/components/account-details/account-details.component.scss
@@ -1,0 +1,91 @@
+.details-layout {
+
+  display: grid;
+  gap: 20px;
+  margin-bottom: 20px;
+
+  .user-image {
+    border: 1px solid #fff;
+    border-radius: 12px;
+    height: 100px;
+    justify-self: center;
+  }
+
+}
+
+.input-container {
+  display: grid;
+  grid-auto-flow: column;
+  grid-auto-columns: max-content;
+  align-items: center;
+  gap: 14px;
+
+  mat-icon {
+    color: var(--color-icon-muted);
+  }
+}
+
+.sign-in-layout {
+  display: grid;
+  gap: 20px;
+
+  .promote-sign-in {
+    color: var(--color-text-muted);
+  }
+
+  ul {
+    margin: 0;
+    padding-left: 1.4em;
+  }
+}
+
+.app-image {
+  height: 100px;
+  justify-self: center;
+}
+
+.email-inputs-container {
+  display: grid;
+  width: max-content;
+  margin: auto;
+}
+
+.link {
+  color: var(--color-hyperlink-text);
+
+  border-radius: 10px;
+  padding: 0 4px;
+  margin: 0;
+  line-height: unset;
+  font-weight: 400;
+  font-size: 0.9em;
+}
+
+.forgot-password {
+  justify-self: end;
+  text-decoration: none;
+}
+
+.sign-in-button-email-container {
+  display: grid;
+  align-items: center;
+  gap: 6px;
+
+  mat-icon {
+    color: var(--color-icon-muted);
+    margin-right: 6px;
+  }
+}
+
+.button-loader {
+  mat-progress-bar {
+    position: absolute;
+    left: 0;
+    bottom: 0;
+  }
+}
+
+.google-g-logo {
+  margin-right: 6px;
+  height: 1.5em;
+}
